chore: ATLAS-666: Convert verification API project to functions project.
