+++ Atlas.MatchPrediction.Test.Verification.Test/Atlas.MatchPrediction.Test.Verification.Test.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="EnumStringValues" Version="4.0.2" />
+    <PackageReference Include="FluentAssertions" Version="5.10.3" />
+    <PackageReference Include="LochNessBuilder" Version="1.0.1" />
+    <PackageReference Include="NSubstitute" Version="4.2.1" />
+    <PackageReference Include="nunit" Version="3.12.0" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Atlas.Common.Test\Atlas.Common.Test.csproj" />
+    <ProjectReference Include="..\Atlas.MatchPrediction.Test.Verification\Atlas.MatchPrediction.Test.Verification.csproj" />
+  </ItemGroup>
+
+</Project>
