+++ Atlas.MatchPrediction.Test.Verification.Test/TestHelpers/NormalisedPoolMemberBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using Atlas.MatchPrediction.Test.Verification.Models;
 using LochNessBuilder;
 
-namespace Atlas.MatchPrediction.Test.Verification.VerificationFrameworkTests.TestHelpers
+namespace Atlas.MatchPrediction.Test.Verification.Test.TestHelpers
 {
     [Builder]
     internal static class NormalisedPoolMemberBuilder
