+++ Atlas.MatchPrediction.Test.Verification.Test/TestHelpers/SimulantBuilder.cs
@@ -2,7 +2,7 @@
 using Atlas.MatchPrediction.Test.Verification.Data.Models;
 using LochNessBuilder;
 
-namespace Atlas.MatchPrediction.Test.Verification.VerificationFrameworkTests.TestHelpers
+namespace Atlas.MatchPrediction.Test.Verification.Test.TestHelpers
 {
     [Builder]
     internal static class SimulantBuilder
