+++ Atlas.MatchPrediction.Test.Verification.Test/TestHelpers/SimulantLocusHlaBuilder.cs
@@ -4,7 +4,7 @@ using Atlas.Common.Test.SharedTestHelpers;
 using Atlas.MatchPrediction.Test.Verification.Models;
 using LochNessBuilder;
 
-namespace Atlas.MatchPrediction.Test.Verification.VerificationFrameworkTests.TestHelpers
+namespace Atlas.MatchPrediction.Test.Verification.Test.TestHelpers
 {
     [Builder]
     internal static class SimulantLocusHlaBuilder
