+++ Atlas.MatchPrediction.Test.Verification.Test/UnitTests/NormalisedPoolGeneratorTests.cs
@@ -1,13 +1,13 @@
-﻿using Atlas.MatchPrediction.Test.Verification.Data.Repositories;
+﻿using System.Linq;
+using System.Threading.Tasks;
+using Atlas.MatchPrediction.Test.Verification.Data.Repositories;
 using Atlas.MatchPrediction.Test.Verification.Services.GenotypeSimulation;
-using Atlas.MatchPrediction.Test.Verification.VerificationFrameworkTests.TestHelpers;
+using Atlas.MatchPrediction.Test.Verification.Test.TestHelpers;
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Atlas.MatchPrediction.Test.Verification.VerificationFrameworkTests.UnitTests
+namespace Atlas.MatchPrediction.Test.Verification.Test.UnitTests
 {
     [TestFixture]
     public class NormalisedPoolGeneratorTests
