+++ Atlas.MatchPrediction.Test.Verification/Functions/HealthCheckFunctions.cs
@@ -1,0 +1,20 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using Atlas.Common.Utils;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
+
+namespace Atlas.MatchPrediction.Test.Verification.Functions
+{
+    public class HealthCheckFunctions
+    {
+        [SuppressMessage(null, SuppressMessage.UnusedParameter, Justification = SuppressMessage.UsedByAzureTrigger)]
+        [FunctionName(nameof(HealthCheck))]
+        public static OkObjectResult HealthCheck([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequest req)
+        {
+            const string responseMessage = "This HTTP triggered function executed successfully";
+            return new OkObjectResult(responseMessage);
+        }
+    }
+}
