+++ Atlas.MatchPrediction.Test.Verification/Functions/SwaggerFunctions.cs
@@ -1,0 +1,31 @@
+﻿using System.Net.Http;
+using AzureFunctions.Extensions.Swashbuckle;
+using AzureFunctions.Extensions.Swashbuckle.Attribute;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
+
+namespace Atlas.MatchPrediction.Test.Verification.Functions
+{
+    public static class SwaggerFunctions
+    {
+        [SwaggerIgnore]
+        [FunctionName(nameof(Swagger))]
+        public static HttpResponseMessage Swagger(
+            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "swagger/json")]
+            HttpRequestMessage req,
+            [SwashBuckleClient] ISwashBuckleClient swashBuckleClient)
+        {
+            return swashBuckleClient.CreateSwaggerDocumentResponse(req);
+        }
+
+        [SwaggerIgnore]
+        [FunctionName(nameof(SwaggerUi))]
+        public static HttpResponseMessage SwaggerUi(
+            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "swagger/ui")]
+            HttpRequestMessage req,
+            [SwashBuckleClient] ISwashBuckleClient swashBuckleClient)
+        {
+            return swashBuckleClient.CreateSwaggerUIResponse(req, "swagger/json");
+        }
+    }
+}
