+++ Atlas.MatchPrediction.Test.Verification/Properties/launchSettings.json
@@ -1,20 +1,8 @@
 ﻿{
-  "$schema": "http://json.schemastore.org/launchsettings.json",
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:30501",
-      "sslPort": 44379
-    }
-  },
   "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      }
+    "Atlas.MatchPrediction.Test.Verification": {
+      "commandName": "Project",
+      "commandLineArgs": "host start --port 7086 --nodeDebugPort 5860"
     }
   }
 }
