+++ Atlas.MatchPrediction.Test.Verification/SwaggerStartup.cs
@@ -1,31 +1,18 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.OpenApi.Models;
+﻿using Atlas.MatchPrediction.Test.Verification;
+using AzureFunctions.Extensions.Swashbuckle;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Hosting;
+using System.Reflection;
+
+[assembly: WebJobsStartup(typeof(SwashBuckleStartup))]
 
 namespace Atlas.MatchPrediction.Test.Verification
 {
-    public static class SwaggerStartup
+    internal class SwashBuckleStartup : IWebJobsStartup
     {
-        public static IServiceCollection ConfigureSwaggerService(this IServiceCollection services)
+        public void Configure(IWebJobsBuilder builder)
         {
-            services.AddSwaggerGen(options =>
-            {
-                options.SwaggerDoc("v1", new OpenApiInfo {Title = "My API", Version = "v1"});
-            });
-            return services;
-        }
-
-        public static IApplicationBuilder ConfigureSwagger(this IApplicationBuilder app, IWebHostEnvironment env = null)
-        {
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
-            {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1");
-                c.RoutePrefix = string.Empty;
-                c.ConfigObject.DisplayRequestDuration = true;
-            });
-            return app;
+            builder.AddSwashBuckle(Assembly.GetExecutingAssembly());
         }
     }
 }
