--- Atlas.MatchPrediction.Test.Verification/appsettings.json
@@ -1,32 +1,0 @@
-{
-  "AllowedHosts": "*",
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "ApplicationInsights": {
-    "InstrumentationKey": "dummy-key",
-    "LogLevel": "Info"
-  },
-  "AzureStorage": {
-    "ConnectionString": "UseDevelopmentStorage=true",
-    "HaplotypeFrequencySetBlobContainer": "haplotype-frequency-set-import"
-  },
-  "ConnectionStrings": {
-    "MatchPredictionVerification:Sql": "Data Source=(local);Initial Catalog=AtlasMatchPredictionVerification;Integrated Security=True;MultipleActiveResultSets=True;",
-    "MatchPrediction:Sql": "Data Source=(local);Initial Catalog=AtlasMatchPrediction;Integrated Security=True;MultipleActiveResultSets=True;"
-  },
-  "HlaMetadataDictionary": {
-    "AzureStorageConnectionString": "UseDevelopmentStorage=true"
-  },
-  "MacDictionary": {
-    "AzureStorageConnectionString": "UseDevelopmentStorage=true",
-    "TableName": "AtlasMultipleAlleleCodes",
-    "Download": {
-      "MacSourceUrl": "https://bioinformatics.bethematchclinical.org/HLA/alpha.v3.zip"
-    }
-  }
-}
